+++ kernel/src/interrupts/keyboard.rs
@@ -26,6 +26,7 @@
 use crate::{
     ports::{self, Port},
     speaker,
+    PANIC,
     startup::{self, ExitCode},
     sysinfo::SystemInfo,
     vga::{
@@ -288,7 +289,7 @@ fn system_command(key: KeyCode, kbd: &Modifiers) {
         match key {
             KeyCode::F1 => print_sysinfo(),
             KeyCode::F3 => speaker::play_song(),
-            KeyCode::F4 => panic!("Triggered System Command 4 by pressing Ctrl+Alt+F4 or SysRq+F4"),
+            KeyCode::F4 => PANIC!(badbug "Triggered System Command 4 by pressing Ctrl+Alt+F4 or SysRq+F4"),
             KeyCode::F5 => super::triple_fault(),
             KeyCode::F6 => buffers::swap(),
             KeyCode::F7 => print_help(),