+++ kernel/src/panic.rs
@@ -19,7 +19,7 @@
 /*!
     kernel/src/panic.rs
 
-    Handles kernel panics created via the [`panic!`] macro
+    Handles kernel panics and the [`PANIC!`] macro.
 */
 
 use crate::{
@@ -28,167 +28,265 @@ use crate::{
     ports::{self, Port},
     speaker,
     sysinfo::SystemInfo,
-    time,
-    vga::{
-        buffers::{self, YoinkedBuffer},
-        cursor::{self, CursorPos},
-        print::Color,
-    },
+    vga::{buffers, cursor},
 };
 use core::{
     arch::asm,
+    ffi::{CStr, c_char, c_void},
     hint,
-    panic::{Location, PanicInfo},
-    sync::atomic::{AtomicBool, Ordering},
+    panic::PanicInfo,
+    sync::atomic::{AtomicU64, Ordering},
 };
 
-/// Stores the stack trace of the last function calls into `frames`.
+/// Sets everything up for, then triggers a kernel panic.
+///
+/// Runs in four different modes, `badbug`, `exception`, `exception noerror`, `const`.
+///
+/// ### badbug
+/// Indicates a general case panic due to a bad bug, such as what would usually cause
+/// the `panic!` macro to be ran.
+/// Takes a format string for the panic's info.
+///
+/// ### exception
+/// Creates a handler function for exceptions, taking the cause of the error as a `&CStr`,
+/// and a `fn(u64)` function pointer to print information about the error code.
+///
+/// ### exception noerror
+/// Same as `exception` except without the error code and it's `fn(u64)` function pointer.
+///
+/// ### const
+/// Const available panicking, taking an `&'static str` as it's cause.
+#[macro_export]
+macro_rules! PANIC {
+    (exception $cause:expr, $info:expr) => {{
+        #[allow(unused_imports)]
+        extern "x86-interrupt" fn wrapper(stackframe: $crate::interrupts::IntStackFrame, errcode: u64) -> ! {
+            use $crate::panic::kpanic;
+            use core::ffi::c_char;
+
+            static mut IP: u64 = 0;
+            static mut ERRCODE: u64 = 0;
+
+            extern "sysv64" fn info() {
+                let errcode = $info;
+
+                // Safety: The statics are only ever written to once
+                unsafe {
+                    let ip = IP;
+                    println!("Instruction: 0x{ip}");
+                    errcode(ERRCODE)
+                }
+            }
+
+            let cause = $cause as *const _ as *const c_char;
+            unsafe {
+                IP = stackframe.ip;
+                ERRCODE = errcode;
+                kpanic(cause, stackframe.sp, info);
+            }
+        }
+
+        wrapper as *const () as u64
+    }};
+
+    (exception noerror $cause:expr) => {{
+        #[allow(unused_imports)]
+        extern "x86-interrupt" fn wrapper(stackframe: $crate::interrupts::IntStackFrame) -> ! {
+            use $crate::panic::kpanic;
+            use core::ffi::c_char;
+
+            static mut IP: u64 = 0;
+            extern "sysv64" fn info() {
+                // Safety: The static's only ever written to once
+                unsafe { let ip = IP; println!("Instruction: 0x{ip}") }
+            }
+
+            let cause = $cause as *const _ as *const c_char;
+            unsafe {
+                IP = stackframe.ip;
+                kpanic(cause, stackframe.sp, info);
+            }
+        }
+
+        wrapper as *const () as u64
+    }};
+
+    (badbug $($fmt:tt)+) => {{
+        let fmt = format_args!($($fmt)+);
+        #[allow(unused)] // may be called from unsafe code or with existing imports
+        unsafe {
+            use core::{ffi::{c_void, c_char}, ptr::null, arch::asm, mem, fmt::Arguments};
+
+            static mut CAUSE:  *const c_char = null();
+            static mut ARGUMENTS: Arguments = format_args!("");
+
+            let rsp: *const c_void;
+            asm!("cli", "mov {0}, rsp", out(reg) rsp);
+
+            // Safety: Since PANIC! never returns, anything local passed to it will live forever
+            ARGUMENTS = mem::transmute::<Arguments<'_>, Arguments<'static>>(fmt);
+            CAUSE = c"BADBUG".as_ptr();
+
+            extern "sysv64" fn info() {
+                // Safety: Only ever written to once above
+                unsafe { let args = ARGUMENTS; println!("{args}") }
+            }
+
+            let cause = CAUSE; // copy out of static
+            asm!(
+                "mov rdi, {0}",
+                "mov rsi, {1}",
+                "mov rdx, {2}",
+                "call kpanic", // must be a call to allow stack trace
+                "jmp hang",
+                in(reg) cause,
+                in(reg) rsp,
+                in(reg) info
+             );
+
+            ::core::hint::unreachable_unchecked()
+        }
+    }};
+
+    (const $cause:expr) => {
+        panic!($cause)
+    }
+}
+
+/// Triggers a kernel panic.
 /// # Safety
-/// The stack trace must be at least `frames` frames deep.
+/// This function should only be called via the [`PANIC`] macro.
+#[rustfmt::skip]
+#[unsafe(no_mangle)]
+pub unsafe extern "sysv64" fn kpanic(
+    cause: *const c_char,
+    sp: *const c_void,
+    info: extern "sysv64" fn(),
+) -> ! {
+    /// The total number of panics which have occurred,
+    /// useful for debugging problems with [`PANIC`] & [`kpanic`].
+    static PANICS: AtomicU64 = AtomicU64::new(0);
+    speaker::stop(); // in case anything was playing, prevent it from playing forever
+    motor::force_disable(); // in case it was on
+    cursor::ALLOW_ROW_0.store(true, Ordering::Relaxed);
+    // Safety: Whoever was using the buffer is long gone now
+    unsafe { buffers::BUFFER_HELD.store(false) };
+
+    // Safety: The caller must ensure that cause points to a valid c str
+    let cause = unsafe { CStr::from_ptr(cause) };
+
+    print!("=============================\n  KERNEL PANIC: ");
+    match cause.to_str() { // remove ugly debug quotation marks if possible
+        Ok(s) => println!("{s}\n"),
+        Err(_) => println!("{cause:?}\n"),
+    }
+
+    // Print kernel & hardware sysinfo
+    let sysinfo = SystemInfo::now();
+    println!(
+        "Kernel: {}{}{}{}{}{} {} {} {} {}",
+        sysinfo.idt_init as u8,
+        sysinfo.gdt_init as u8,
+        sysinfo.pic_init as u8,
+        sysinfo.pit_init as u8,
+        sysinfo.kbd_init as u8,
+        sysinfo.fdc_init as u8,
+        sysinfo.time,
+        sysinfo.debug as u8,
+        PANICS.fetch_add(1, Ordering::Relaxed),
+        sysinfo.sfk_version_short
+    );
+    print!(
+        "Hardware: {} {} ",
+        sysinfo.cpu_vendor,
+        sysinfo.floppy_space.unwrap_or(&0)
+    );
+    match sysinfo.date { // print the date if we have it
+        Ok(d) => println!("{d}\n"),
+        Err(e) => println!("{}\n", e.state),
+    };
+
+    info();
+    stack_trace(6);
+
+    // Print the top few elements on the stack
+    // Safety: PANIC should have (hopefully) sent through a valid SP
+    let valof = |offset| unsafe { *((sp as *const u64).wrapping_add(offset)) };
+    println!(
+        "\nStack (SP=0x{sp:?}):\n  {:#18x}  {:#18x}  {:#18x}\n  {:#18x}  {:#18x}  {:#18x}",
+        valof(0),
+        valof(1),
+        valof(2),
+        valof(3),
+        valof(4),
+        valof(5),
+    );
+
+    // Loop waiting for kbd input
+    print!("\nPress ESC to restart device");
+    cursor::update_visual_pos();
+    loop {
+        const ESC_SCANCODE_SET1: u8 = 0x01;
+        const ESC_SCANCODE_SET2: u8 = 0x76;
+        // Safety: Port 0x60 is fine to read as it just contains the last scancode
+        let scancode = unsafe { ports::readb(Port::PS2Data) };
+        if scancode == ESC_SCANCODE_SET1 || scancode == ESC_SCANCODE_SET2 {
+            interrupts::triple_fault()
+        }
+        hint::spin_loop() // can't halt because of cli
+    }
+}
+
+/// Prints a stack trace at most `frames` stackframes up.
 #[unsafe(no_mangle)]
 #[inline(never)]
-unsafe fn stack_trace(frames: &mut [u64]) -> usize {
+fn stack_trace(frames: u32) {
     #[repr(C)]
+    #[derive(Clone, Copy)]
     struct Stackframe {
         next: *const Stackframe,
         rip: u64,
     }
 
-    let rbp: *const *const Stackframe;
-    // Safety: RBP should hopefully be a ptr to the function's stackframe
-    let mut stack = unsafe {
-        asm!("mov {0}, rbp", out(reg) rbp);
-        *rbp
-    };
+    let mut stack: *const Stackframe;
+    // Safety: RBP (should) always point to the last stackframe,
+    // even after interrupt handlers have been fired
+    unsafe { asm!("mov {0}, rbp", out(reg) stack) }
 
-    for frame in frames.iter_mut() {
-        // Safety: The caller must ensure that the trace is contained within '0..frames'
-        let rip = unsafe { *(stack.wrapping_byte_add(8) as *const u64) };
-        if !(0x209000..=0x220000).contains(&rip) || !stack.is_aligned() {
-            // return early if the stack isn't in a 'safe' range or the deref will fail
-            return rbp.addr();
+    println!("\nStack trace (BP=0x{stack:?}):");
+    for idx in 0..frames {
+        // Safety: See safety comment above
+        let sf = unsafe { *stack };
+        stack = sf.next;
+
+        // bootloader nicely ends the stackframe list with a null for us
+        if stack.is_null() {
+            return;
         }
 
-        *frame = rip;
-        // Safety: At least we know that the addr is in the right space & aligned
-        stack = unsafe { (*stack).next };
+        if sf.rip != 0 {
+            println!("  {idx}  {:#8x}", sf.rip)
+        }
     }
-
-    rbp.addr()
 }
 
-/// Ran when a kernel panic occurs.
+/// Ran when the `panic!` macro is invoked.
 #[panic_handler]
 #[cfg_attr(test, allow(unused))]
-fn kernel_panic(info: &PanicInfo) -> ! {
+fn panic_handler(panic_info: &PanicInfo) -> ! {
     #[cfg(test)]
     {
         // tests fail by panicking
         use crate::tests::exit_qemu;
-        println!("- failed, see failure cause below\n{info}");
+        println!("- failed, see failure cause below\n{panic_info}");
         exit_qemu(true);
     }
 
-    // !!!!!!!!!
-    interrupts::cli();
-    time::set_waiting_char(false);
-    time::WAITING_CHAR.store(false, Ordering::Relaxed);
-    speaker::stop(); // in case anything was playing, prevent it from playing forever
-    motor::force_disable(); // in case it was on
-    // Safety: Whoever was using the buffer is long gone now
-    unsafe { buffers::BUFFER_HELD.store(false) };
-
-    // Swap & wipe screen
-    cursor::ALLOW_ROW_0.store(true, Ordering::Relaxed);
-    buffers::swap();
-    buffers::clear();
-
-    let location = info.location().unwrap(); // always succeeds
-    let sysinfo = SystemInfo::now();
-    println!(
-        fg = Grey,
-        "                                  KERNEL PANIC\n
-      Sunflower encountered a kernel panic at {}:{}:{}\n
-      System information: {} | {} | {} | {} | {}\n
-      Press ESC to restart device and ENTER to show previous screen\n
-                           Press any key to continue\n
-      Error information:\n
-      {}\n",
-        location.file().trim_prefix("src/"),
+    let location = panic_info.location().unwrap(); // always succeeds
+    PANIC!(badbug "External panic at {}:{}:{}\n{}",
+        location.file(),
         location.line(),
         location.column(),
-        sysinfo.sfk_version_short,
-        sysinfo.cpu_vendor,
-        sysinfo.debug as u8,
-        sysinfo.time,
-        sysinfo.floppy_space.copied().unwrap_or_default(),
-        info.message()
-    );
-
-    let mut buf = [0; 6];
-    // Safety: The return early check should hopefully ensure nothing bad happens
-    let rbp = unsafe { stack_trace(&mut buf) };
-    println!(fg = Grey, "      Stack frame (RBP=0x{rbp:x}):");
-
-    for (idx, frame) in buf.iter().filter(|f| **f != 0).enumerate() {
-        println!(fg = Grey, "         {idx}   {frame:#8x}")
-    }
-
-    paint_screen();
-    loop {
-        check_keyboard(location);
-        hint::spin_loop(); // can't halt because of cli
-    }
-}
-
-/// Paints the screen's background color.
-fn paint_screen() {
-    /// The background color to paint the screen.
-    const PANIC_COLOR: Color = Color::Red;
-
-    // Yoink always succeeds
-    if let Some(mut buf) = YoinkedBuffer::try_yoink() {
-        for row in buf.buffer().iter_mut() {
-            for px in row {
-                let px = px.as_raw_mut();
-                *px &= !(0b1111 << 12); // clear bg
-                *px |= (PANIC_COLOR as u16) << 12; // set bg
-            }
-        }
-    }
-
-    // Move cursor to end of press any key line
-    CursorPos::set_row(11);
-    CursorPos::set_col(52);
-    cursor::update_visual_pos();
-}
-
-/// Triple faults if `ESC` is pressed & prints sysinfo if `ALT` is pressed.
-fn check_keyboard(location: &Location) {
-    /// Should we allow swapping the buffers?
-    static ALLOW_BUFSWAP: AtomicBool = AtomicBool::new(true);
-
-    /// Scancodes in set 2.
-    const ESC_SCANCODE: u8 = 0x76;
-    const ENTER_SCANCODE: u8 = 0x5A;
-
-    // Safety: Port 0x60 is fine to read as it just contains the last scancode.
-    let scancode = unsafe { ports::readb(Port::PS2Data) };
-    if scancode == ESC_SCANCODE {
-        interrupts::triple_fault();
-    } else if scancode == ENTER_SCANCODE && ALLOW_BUFSWAP.fetch_and(false, Ordering::Relaxed) {
-        buffers::swap();
-        CursorPos::set_row(u8::MAX);
-        CursorPos::set_col(0);
-        cursor::update_visual_pos();
-        print!(
-            fg = Grey,
-            "-------------------------------------------------------------------------------- {} panicked at {}:{}:{} | Press ESC to restart",
-            env!("SFK_VERSION_SHORT"),
-            location.file().trim_prefix("src/"),
-            location.line(),
-            location.column()
-        );
-        cursor::update_visual_pos();
-    }
+        panic_info.message()
+    )
 }